move tokenizer tests to their own test file
